+++ src/main/java/com/example/mogakserver/auth/api/request/SignUpRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.mogakserver.auth.application.request;
+package com.example.mogakserver.auth.api.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
