+++ src/main/java/com/example/mogakserver/auth/api/request/TokenRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.mogakserver.auth.application.request;
+package com.example.mogakserver.auth.api.request;
 
 public record TokenRequestDto(String accessToken, String refreshToken) {
 }
